services/orchestrator/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
google-cloud-firestore==2.16.0
httpx==0.27.2
